L2Syncer enhancements
