+++ crates/common/src/lib.rs
@@ -5,7 +5,6 @@ pub mod backup;
 pub mod cache;
 pub mod config;
 pub mod da;
-pub mod db_migrations;
 pub mod l2;
 pub mod rpc;
 pub mod utils;
